requirements.txt
# Core - only what you need
streamlit
python-dotenv

# PDF processing
pypdf

# Embeddings (much smaller alternative)
sentence-transformers

# Vector DB
pinecone

# LLM
groq

# Text processing
langchain-text-splitters
langchain-community
